Move encoding to module-lattice module
